+++ changelog.md
@@ -3,6 +3,12 @@ All notable changes to WiNER-fr will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [WiNER v0.2.0](https://github.com/YoannDupont/WiNER-fr/releases/tag/v0.2.0)
+### Added
+- python script to gather wikinews articles.
+- python script to get some numbers about the corpus.
+- every missing month from Wikinews 2018.
+
 ## [WiNER v0.1.0](https://github.com/YoannDupont/WiNER-fr/releases/tag/v0.1.0)
 ### Added
 - python script to detect some inconsistencies.